Fix zsh vi mode.
Oh, My Zsh! was resetting my bindings. I need to tell Oh, My Zsh! to
load the `vi-mode` plugin. I'm now specifying the plugins to load in the
`ohmy.zsh` configuration and commenting out the plugin specification
line of the generated `.zshrc` file.

This and a general reordering of the `zprofile.d` directory.