Save for posterity.
Created a backup strategy using Duplicity that only adds or removes
files, but does not save meta data changes. If a file in the volume is
edited, it will not be saved until another file is added or removed from
the volume. This is for volumes that accumulate data, not for volumes
where work is performed.

See: http://goo.gl/otlZv