Use Duplicity output for dirty diffs.
To determine if a posterity volume is dirty, use the relative path
output by `duplicity` instead of making an effort to trim the paths so
that they are relative to the volume directory.

Also, no more pruning the listings of `.DS_Store` and `.AppleDouble`.
Duplicity is configured to ignore these files, so we'll depend on
Duplicity's listings of changes and volume contents.