Copy dist tag info from temp file.
A Vulcan nerve pinch on a running `dots node outdated` will leave an
empty dist tag info file in the cache. That will cause `dots node
outdated` to report unhappiness until the cache entry expires. We now
download to a temp file and then copy that file which seems to be less
likely to cause problems even though copy is not atomic. It is close
enough to atomic because files are so small.