`node recurse` execute and visit separated.
The `node recurse` root option will run the command against the root of
a collection of qualified packages. It appeared that once it ran the
root program, when it encountered a subsequent module that was also from
the same collection it would skip it because it had been marked visited.

We now mark modules or roots executed in a separate set from modules
visited.