Return on last error in argument parser.
We emit our `set --` after the last error encountered by the argument
parser so that if the user defined error handler did not emit an error
exit processing of the arguments can continue. Essentially a way to stop
the parser on the first argument it does not understand.

This was thought to be necessary to support shebang arguments, where the
file name given to use by `/usr/bin/env` has no switches, but the
swapped function is expecting switches.

We're probably going to just do some intense peeking.

Note that if we switched to contemporary argument handling, where you
can intersperse switched arguments with positional arguments, this might
not be necessary, and our argument parser could handle this case.