Let's expore Emacs for Markdown.
Definately further along than ViM or Neovim. Able to collapse and show a
readable view. My light-mode colors are currently terrible for this
though, so I'd have to get color sorted out so that it would look near
as nice as Goyo.

To get Goyo back you'll need one of the following.

 * https://github.com/joostkremers/writeroom-mode
 * https://github.com/rnkn/olivetti
 * https://github.com/joaotavora/darkroom

Note that you're following the instructions here.

 * https://leanpub.com/markdown-mode/read

You're going to have to learn the Emacs key bindings, or else you're
going to have to set vi keybindings in emacs and hope that you're still
able to use the Markdown features.

I'm committing just the part that allowed for intalling the Markdown
mode. Use the Emacs installer to install it and not Homebrew.