Learn ViM the Hard Way, fix Goyo.
Got Goyo to work again. Limelight is currently broken, though, so I
disabled it so that Goyo would initialize with Pencil.

Started working through Learn ViM the Hard Way, so there are some
scribbles in `etc/vimrc` that part part of the exercises. Checking them
in so as to continue soon, I am hoping.